slide_builder.py
from pptx import Presentation
from pptx.util import Inches, Pt
from pptx.dml.color import RGBColor
from pptx.enum.text import PP_ALIGN

def generate_pptx(plan, image_paths, output_path="presentation.pptx"):
    """
    Generates PPTX based on the AI plan and generated images.
    """
    prs = Presentation()
    
    # Extract theme
    theme = plan.get('theme', {})
    font_name = theme.get('font', 'Arial')
    primary_color_hex = theme.get('primary_color', '#000000').lstrip('#')
    
    # 1. Title Slide (Slide 0 in plan?, usually we generate a cover manually or it's first in slides)
    # The parser puts the main title in 'title' field of parsed data, but plan has 'slides'
    # Let's assume the user wants the first slide in 'slides' to be content. 
    # But wait, Markdown parser extracted a global "Title".
    # Usually we want a Cover Slide.
    
    # Let's check if the Plan includes a Cover. 
    # Current ai_planner prompts for "slides" based on content.
    # We should add a generic Title Slide first.
    
    # slide_layout 0: Title Slide
    slide = prs.slides.add_slide(prs.slide_layouts[0])
    title = slide.shapes.title
    title.text = "Presentation" # We need the main title passed here or in plan
    # The plan structure in my ai_planner doesn't explicitly carry the Top Level Title.
    # I should pass it. 
    
    # ADJUSTMENT: Let's assume the calling 'app.py' passes the main title or it's implicitly handled.
    # For now, I'll allow a "title" argument or just take the first slide.
    
    # Clean output slides
    # Let's iterate through the PLAN's slides.
    
    for i, slide_data in enumerate(plan['slides']):
        layout_type = slide_data.get('layout', 'title_bullets')
        
        # Select Layout
        if layout_type == 'hero_image':
             # Blank layout, manual placement
             slide = prs.slides.add_slide(prs.slide_layouts[6]) 
        elif layout_type == 'title_bullets_image_right':
            # Title and Content (we will resize content)
            slide = prs.slides.add_slide(prs.slide_layouts[1])
        else:
            # Standard Title and Content
            slide = prs.slides.add_slide(prs.slide_layouts[1])
            
        # --- Content Generation ---
        
        # 1. Images
        # Check if we have an image for this slide index
        image_path = image_paths.get(i)
        
        if layout_type == 'hero_image' and image_path:
            # Full screen image
            left = top = Inches(0)
            pic = slide.shapes.add_picture(image_path, left, top, height=prs.slide_height)
            # Center horizontally if aspect ratio differs
            if pic.width < prs.slide_width:
                 pic.left = (prs.slide_width - pic.width) // 2
            
            # Overlay Title
            txBox = slide.shapes.add_textbox(Inches(1), Inches(3), Inches(8), Inches(2))
            tf = txBox.text_frame
            p = tf.add_paragraph()
            p.text = slide_data['title']
            p.font.size = Pt(54)
            p.font.color.rgb = RGBColor(255, 255, 255)
            p.font.bold = True
            p.alignment = PP_ALIGN.CENTER
            
        elif layout_type == 'title_bullets_image_right':
            # Title
            slide.shapes.title.text = slide_data['title']
            
            # Bullets (Left side)
            # Resize placeholder
            body = slide.shapes.placeholders[1]
            body.left = Inches(0.5)
            body.width = Inches(4.5) # Half width roughly
            
            tf = body.text_frame
            tf.clear() # Clear default empty paragraph
            for bullet in slide_data['bullets']:
                p = tf.add_paragraph()
                p.text = bullet
                p.level = 0
                p.font.size = Pt(20)

            # Image (Right side)
            if image_path:
                left = Inches(5.2)
                top = Inches(2)
                # Height constraint
                height = Inches(4) 
                slide.shapes.add_picture(image_path, left, top, height=height)
                
        else: # Standard or Title Only
            slide.shapes.title.text = slide_data['title']
            
            if 'bullets' in slide_data and slide_data['bullets']:
                tf = slide.shapes.placeholders[1].text_frame
                tf.clear()
                for bullet in slide_data['bullets']:
                    p = tf.add_paragraph()
                    p.text = bullet
                    p.level = 0
    
    prs.save(output_path)
    return output_path
