requirements.txt
streamlit
python-pptx
google-generativeai
Pillow
python-dotenv
